Project
In this project i am able to develope a login page using the mysql database and use it for checking if the student is registered to the respective mess.
I have used the timestamp option in the table to save the time of login.
I was trying to build a incrementing variable to store the number of logins
with differenciation(breakfast/luch/dinner)
At the last in the billing section i stored the number of times a student has logged in to create the dues towards the respective Mess